review

+++ wrappers/python/tests/layer/test_gpt2_attention.py
@@ -33,7 +33,7 @@ dtype2nntile = {
 }
 
 dtype2np = {
-    'fp16': np.float16,
+    'fp16': np.float32,
     'bf16': np.float16,
     'fp32': np.float32,
 }
