+++ wrappers/python/tests/model/test_gpt2_model.py
@@ -39,7 +39,7 @@ dtype2nntile = {
 dtype2np = {
         'fp32': np.float32,
         'bf16': np.float16,
-        'fp16': np.float16,
+        'fp16': np.float32,
 }
 
 dtype2tol = {
